// src/types.d.ts
// Generated by https://quicktype.io

export interface CredValidation {
    canLoginWhenInTestMode: boolean;
    emailAddress: null;
    inTestModeMessage: null;
    isClassicSite: boolean;
    loginStatus: number;
    redirectUrl: string;
    schoolName: string;
    studentApiUrl: string;
    userId: number;
    userType: number;
    validationId: string;
    validationToken: string;
}

// Generated by https://quicktype.io

export interface StudentCredValidation {
    authToken: string;
    firstName: string;
    lastName: string;
    nickName: string;
}


// Generated by https://quicktype.io

export interface ScheduleView {
    appointmentDate: string;
    appointmentId: number;
    courseId: number;
    courseName: null | string;
    courseRoom: null | string;
    isQuickOptionAvailable: boolean;
    meetingLink: null;
    periodDescription: string;
    periodId: number;
    schedulerComment: null;
    schoolIsBlockedForDatePeriod: boolean;
    studentCanOverwrite: boolean;
}

// Generated by https://quicktype.io

export interface ScheduleListResponse {
    additionalMessage: null;
    courses: Course[];
    courseListStatusId: number;
    currentlyScheduledCourseId: number;
    schedulerComment: null;
    wasFoundInCache: boolean;
}

export interface Course {
    appointmentRequestCourseId: number;
    appointmentTypeId: number;
    blockedDateDetailId: number;
    blockedDateId: number;
    blockedReason: null;
    courseNameOriginal: string;
    currentCourseId: number;
    isAdjusted: boolean;
    isOpenForScheduling: boolean;
    blockedType: number;
    courseDescription: string;
    courseId: number;
    courseRoom: string;
    defaultAppointmentTypeId: number;
    departmentId: number;
    departmentName: string;
    maxNumberStudents: number;
    periodId: number;
    preventStudentRequesting: boolean;
    preventStudentSelfScheduling: boolean;
    stafferFirstName: string;
    stafferId: number;
    stafferLastName: string;
    appointmentDate: string;
    schoolId: number;
    numberOfAppointments: number;
}